+++ src/dash/DefaultDash.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import DashSearchBar from './DashSearchBar.jsx';
-import DefaultDashTutorList from './TutorSearchResult.jsx';
+import DefaultDashTutorList from './TutorSearchList.jsx';
 
 const tutors = [
     {
